projects/DocuSwarm/archived_projects/DocuSwarm/requirements.txt
Flask==3.0.0
google-generativeai
python-dotenv
gunicorn
psycopg2-binary
clerk-backend-api
stripe
PyJWT[crypto]
requests
posthog
